feat - restructured the app structure
