+++ app/models/__init__.py
@@ -1,0 +1,13 @@
+# import glob
+# from os.path import basename
+# from os.path import dirname
+# from os.path import isfile
+# from os.path import join
+
+# modules = glob.glob(join(dirname(__file__), "*.py"))
+
+# __all__ = [basename(f)[:-3] for f in modules if isfile(f) and not f.endswith("__init__.py")]
+
+from .users import User
+
+__all__ = ["User"]
