+++ app/routes/__init__.py
@@ -1,0 +1,3 @@
+from .users import user
+
+__all__ = ["user"]
