+++ scripts/local_server.sh
@@ -1,2 +1,2 @@
 # Run migrations and start server
-alembic upgrade head && uvicorn app:app --host 0.0.0.0 --port 8000
+alembic upgrade head && uvicorn app.app:app --host 0.0.0.0 --port 8000
